decouple DescribeCommandReporter
